HoshinoBot/hoshino/modules/chachengfen/requirements.txt
aiofiles>=0.8.0
httpx>=0.22.0
Jinja2>=3.0.3
loguru>=0.6.0
Pillow>=9.1.0
playwright>=1.20.1
